+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.1.8]
+
+- Update toast for successful swap and add abort controller to fetch assets function.
+
 ## [1.1.7]
 
 - Fixed USD value formatting and rate display.
@@ -20,7 +24,6 @@
 - Updated balance retrieval function to use fallback RPC clients as default instead of default wallet RPCs.
 - Rate conversion and display estimated token price for inputAsset.
 - Fixed fiat token prices.
-
 
 ## [1.1.4]